refactor: 키보드 컨트롤러 리팩토링 - 기능 중복하는 코드 삭제 - navigateToElements -> moveToNextFocus로 개명 - moveToNextFocus는 기능 수행만 있어서 핸들러로 분리 - 핸들러에서 focus 처리하면 의존성 중복 -> 컨트롤러가 처리하게 변경
